Compute a Citations Network of functions form a corpus_list dataframe

This function read a standard list of data.frame (class corpus.list) select the codes data.frame and append a Citations Network to the corpus (see hereafter). The classes of this new entry are data.frame citations.network

Usage

compute_doc_network_from_corpus(
  corpus,
  prefix_for_2nd_matches = "\\b",
  suffix_for_2nd_matches = "\\(",
  filter_egolink_within_a_file = TRUE,
  file_path_from_colname = "from",
  file_path_to_colname = "to",
  function_matched_colname = "function",
  line_number_matched_colname = "line_number",
  content_matched_colname = "content",
  order_fn_called_colname = "order"
)

Arguments

corpus

character A corpus.list object from the construct_corpus function

prefix_for_2nd_matches

character A string representing the prefix to add to each 1st match that will be turned into a new regular expressions. The default is an empty string.

suffix_for_2nd_matches

character A string representing a regex to add as a suffix of each match, in order to have a complete regular expression. The default is an empty string.

filter_egolink_within_a_file

logical, default = TRUE. A logical value indicating whether to filter results based on "ego links" (a document referring to itself)

file_path_from_colname

character, default = 'from' The column name (as a string) representing the "from" file path in the result data frame.

file_path_to_colname

character, default = 'to' The column name (as a string) representing the "to" file path in the result data frame.

function_matched_colname

character, default = 'function' The column name for the matched text (default is a function name) in the result data frame.

line_number_matched_colname

character, default = 'line_number' The column name (as a string) for the line number of the second match in the result data frame.

content_matched_colname

character, default = 'content_matched' The column name (as a string) for the full line where a match have occurred.

order_fn_called_colname

character, default = 'order_call' The order of the functions called in a file, based on line number.

Value

A list of dataframe symbolizing the edgelist of a document-to-document citations network

from

character Citations Network - The local file path or GitHub URL that call a function.

to

character Citations Network - The local file path or constructed GitHub URL where the function called is defined.

file_path

character Corpus - The local file path or constructed GitHub URL.

line_number

integer Corpus - The line number within the file.

content

character The content where a match could occur (citations.network), within a file (corpus.nodelist) or a line (corpus.lines).

match

character Corpus - The matched text on this line, NA if there is no match.

Details

It is designed to generate a network of text by cascading text research, assuming the 1st matches are already realized by construct_corpus :

The function will craft a pattern by appending all the 1st matches (matches column elements) , eventually adding a suffix and a prefix to these elements.

Then it will perform a direct extraction with this pattern, and return the corpus with a new data.frame of class citations.network (document with the 2nd match => document with the 1st match). By default, egolinks are removed since filter_egolink_within_a_file default is TRUE

See also

Examples

{
# Example with local folder path
corpus <- construct_corpus(folders =  "~", languages = "R")
corpus <- compute_doc_network_from_corpus(corpus)
# Return a list of df (1st one is supposed to be an edgelist)
# (from the file where a function is call => to the file were defined)
}
#> Error in sub(re, "", x, perl = TRUE): input string 8 is invalid UTF-8