Get lines text-metrics from a df

Given a character vector, return a data.frame with lines text-mining metrics : values are related to each element of the vector provided by the user.

Usage

compute_nchar_metrics(
  text,
  nchar_colname = "n_char",
  nchar_nospace_colname = "n_char_wo_space",
  word_count_colname = "n_word",
  vowel_count_colname = "n_vowel"
)

Arguments

text

character. A character vector of texts.

nchar_colname

character. Default = ""n_char"". character indicating the colname for total nchar().

nchar_nospace_colname

character. Default = "n_char_wo_space". character indicating the colname for nchar() without space.

word_count_colname

character. Default = "n_word". character indicating the colname for a word count.

vowel_count_colname

character. Default = "n_vowel". character indicating the colname for a vowel count (as a syllab count).

Value

A data.frame sorted in the same order as the character vector provided by the user, providing for each element of the vector :

n_char

character Number of characters - including spacing - in a line.

n_char_wo_space

character Number of characters - without spacing - in a line.

word_count_colname

character Number of words in a line.

vowel_count_colname

character Number of vowels in a line.

Examples

if (FALSE) { # \dontrun{
# Example 1: Construct a corpus from local folders
 texts = c("the lazy rat terrier jump over the quick brown fox"
 , "cause lazy rat terrier are the best dog")
 compute_nchar_metrics(texts)
} # }