Construct a Data Frame of Code Files from Local GitHub Repositories and/or Local Folders

This function builds a structured data frame containing code files from either GitHub repositories or local folder paths. Then it try to extract function names in the lines previously readed , according to the default pattern designed for matching functions names. The resulting data frame provides an overview of the collected content, including file paths, URLs (if applicable), and languages.

Usage

construct_corpus(
  local_folders_paths = NULL,
  languages = "R",
  repos = NULL,
  pattern_to_exclude_path = NULL,
  autoclean_filespath = TRUE,
  ignore_match_less_than_nchar = 3,
  pattern_to_remove = "https://raw.githubusercontent.com/",
  ...
)

Arguments

local_folders_paths

character. Default = NULL. A character vector of local folder paths to scan for code files.

languages

character. Default = "R". A character vector specifying the programming language(s) to include in the corpus.

repos

character. Default = NULL. A character vector of GitHub repository URLs or repository identifiers to extract files from (e.g., "user/repo").

pattern_to_exclude_path

character. A regular expression pattern to exclude files by scanning their full names (path/filename.ext). Default is NULL (no files are excluded).

autoclean_filespath

logical. Default = TRUE. Codexplor try to exclude testing files before to read the files of a project ; and it will try to clean the urls of the returned df By default, typical testing-purpose .R files are excluded (i.e. "\.Rcheck|test-|Vignette") , in addition to an optional pattern_to_exclude_path passed by the user.

ignore_match_less_than_nchar

integer. Default = 3. Number of characters for the 1st match to be considered valid.

pattern_to_remove

character Default = 'https://raw.githubusercontent.com/' A pattern (regex) to remove from the files path of the edgelist (columns 1 & 2)

...

Additional arguments passed to srch_pattern_in_files_get_df (filtering options, depth of folder scanning, names of the returned df columns, .verbose parameter, etc.).

Value

A data frame containing the corpus of collected files. The data frame typically includes columns such as:

file_path

character The local file path or constructed GitHub URL.

line_number

integer The local file path or constructed GitHub URL.

content

character The local file path or constructed GitHub URL.

n_char_line

integer Total nchar of a line, including spacing.

n_char_wo_space_line

integer Total nchar of a line, without spacing.

nchar_wo_space_in_file

integer Sum of n_char_wo_space_line for a given file path (same value is repeated for each file path)

Details

If local_folders_paths is provided (one or a list), the function scans the directories and retrieves file paths matching the specified languages.

If repos is provided (one or a list), it constructs URLs to the raw content of files from the specified GitHub repositories.

Both local paths and GitHub URLs can be combined in the final output.

See also

Examples

# Example 1: Construct a corpus from local folders
cr1 <- construct_corpus("~", languages = c("R", "Python"), .verbose = FALSE)
#> Error in sub(re, "", x, perl = TRUE): input string 8 is invalid UTF-8
if (FALSE) { # \dontrun{
# Example 2: Construct a corpus from GitHub repositories
cr2 <- construct_corpus(repos = c("tidyverse/stringr", "tidyverse/readr"), .verbose = FALSE)

# Example 3: Combine local folders and GitHub repositories
cr3 <- construct_corpus("~", c("R", "Cpp"), "tidyverse/dplyr", .verbose = FALSE)
} # }