Create a corpus.list from a list or append a dataframe to a corpus.list

Create a corpus.list from a list or append a dataframe to a corpus.list

Usage

.construct.corpus.list(
  corpus,
  df_to_add = NULL,
  names_of_df_to_add = NULL,
  ...
)

Arguments

corpus

corpus.list list of dataframes created by construct_corpus().

df_to_add

list of data.frame or a single data.frame to append to the corpus.list. If data.frame(s) without a name are provided, names should be provided with names_of_df_to_add

names_of_df_to_add

character vector of names to append to the corpus.list, if anonymous materials is provided (unamed df or list of unamed df).

...

Additionnal attributes (e.g., keep trace of the languages analyzed).

Value

A dataframe of class 'corpus.list'.