Extract Nested Blocks and Assign Group IDs and Deep Levels

Assuming the user have passed a character vector, this function extracts nested blocks of text (e.g., Within (Parenth.)), delimited by a specified opening and a closing characters. It returns a data frame with :

the index of the input string ;

the extracted block (including delimiters) ;

a group_id common to all blocks nested within the same outer block;

a deep-level numeric indicator, within the longest block

Usage

extract_nested_blocks_from_text(
  text,
  delims = c(`(` = ")", `[` = "]", `{` = "}"),
  return_text_col = TRUE
)

Arguments

text

character vector containing one or more text(s)

delims

list of characters symbolizing opening and closing characters, default = c("(" = ")", "[" = "]", "{" = "}").

return_text_col

logical, default = FALSE Return a text column if set to TRUE

Value

A data frame with 4 or 5 columns :

instruction_id

Numeric (row_number)

block

Character (text extracted, surrounded by delimiters)

level

Numeric (level of imbrication of a block surrounded by the delimiters, from 1 to Inf)

group_id.

Numeric, symbolize the imbrication within a same block, i.e. surrounded by global delimiters)

text

Original characters vector passed by the user (optionnally returned)

Examples

text <- "Voici un (exemple (avec un texte (imbriqué)) et [un autre exemple]). MAIS ENSUITE (ceci)"
extract_nested_blocks_from_text(text)
#>                                                        block level group_id
#> 4 (exemple (avec un texte (imbriqué)) et [un autre exemple])     1        1
#> 2                                 (avec un texte (imbriqué))     2        1
#> 3                                         [un autre exemple]     2        1
#> 1                                                 (imbriqué)     3        1
#> 5                                                     (ceci)     1        2
#>                                                                                       text
#> 4 Voici un (exemple (avec un texte (imbriqué)) et [un autre exemple]). MAIS ENSUITE (ceci)
#> 2 Voici un (exemple (avec un texte (imbriqué)) et [un autre exemple]). MAIS ENSUITE (ceci)
#> 3 Voici un (exemple (avec un texte (imbriqué)) et [un autre exemple]). MAIS ENSUITE (ceci)
#> 1 Voici un (exemple (avec un texte (imbriqué)) et [un autre exemple]). MAIS ENSUITE (ceci)
#> 5 Voici un (exemple (avec un texte (imbriqué)) et [un autre exemple]). MAIS ENSUITE (ceci)