Retrieve function definition regex patterns by programming language(s)

This function returns a df with informations for each language (each row), e.g., file extensions associated and regex pattern for matching commented lines and functions definitions) for one or more specified programming languages.

Usage

get_def_regex_by_language(languages = NULL, ...)

Arguments

languages

character, default = "R" A programming language to match (character strings).

...

character One or more programming language names (character strings).

Value

A list where each entry corresponds to a language and contains:

language

character The name of the language on a row, e.g., 'R'.

fn_regex

character A regex dedicated to catch the function names, as soon as a function is defined within a file.

file_extension

character The typical programming file extension, e.g., ".R" for the R language.

commented_line_char

character The pattern symbolizing a commented line, i.e. content is commented after that pattern.

delim_pair

character list indicating the - open and close - characters that symbolizes a multi-lines comment, in addition to the commented_line_char one-liner syntax.

pattern_to_exclude

character The pattern of typical programming files to exclude from the analyses, e.g., "\\.Rcheck|test-|vignettes" for the R language.} \item{\code{local_file_ext}}{characterThe typical programming file extension turned into a regex, by pasting"$"to the end offile_extension` value.

Details

This function supports multiple languages in a single call. Language names are case-insensitive.

Examples

fn_def <- get_def_regex_by_language("Python", "R" , "Go", "C", "JavaScript")
#> Warning: Unsupported language(s) (go) !
#>  Available languages => R, Python, JavaScript, Java, C, Cpp
names(fn_def) ; str(fn_def[[1]])
#> [1] "R"          "Python"     "JavaScript" "C"         
#> List of 6
#>  $ fn_regex           : chr "(^| \\.|\\b)([A-Za-z0-9_\\.]+)(?=\\s*(?:<-)\\s*function)"
#>  $ file_extension     : chr ".R"
#>  $ commented_line_char: chr "\\s?#"
#>  $ delim_pair         : logi NA
#>  $ pattern_to_exclude : chr "\\.Rcheck|test-|vignettes"
#>  $ local_file_ext     : chr ".R$"