Turn a programing project into a Documents Network

This function craft a corpus, according to the default settings. It will return the corpus with a citations network of internal dependancies

Usage

get_doc_network_from_project(
  folders = NULL,
  repos = NULL,
  languages = "R",
  ...
)

Arguments

folders

character A string or list representing the path(s) of local folders path to read.

repos

character A string or list representing the name(s) of github repos (e.g., 'tidyverse/stringr').

languages

character. Default = "R". A character vector specifying the programming language(s) to include in the corpus.

...

Parameters passed to add_doc_network_to_corpus(). These parameters are

characters values, in order to add a prefix and a suffix to the pattern searched (e.g., suffix_for_2nd_matches) or changing the colnames (e.g., file_path_from_colname).

logical values, e.g., "filter_egolink_within_a_file" (default = TRUE)

Value

A list of 4 dataframe : 2 of class corpus.lines, 1 corpus.nodelist and 1 citations.network (symbolizing the edgelist of a document-to-document citations network within a programming project)

from

character citations.network - The local file path or GitHub URL that call a function.

to

character citations.network - The local file path or constructed GitHub URL where the function called is defined.

function

character citations.network - The name of the function matched on a line.

content_matched

character citations.network - The full content matched with the 2nd matches, in order to verify and craft a new regex.

line_number

character citations.network & corpus.lines - The line number of the 2nd match (citation.network) or associated with a line (corpus.lines).

file_path

character corpus.lines & corpus.nodelist - The local file path or constructed GitHub URL, same values as the from & to columns of the citations.network df.

content

character corpus.lines - The content from a line.

matches

character corpus.lines (specifically the codes data.frame)

The matched text during the 1st matches (full of NA if there is no match or if they are filtered out, the default).

See also

Examples

# Example with url from github
result <- get_doc_network_from_project(folder =  "~" )
#> Error in sub(re, "", x, perl = TRUE): input string 8 is invalid UTF-8
# Return a list of df
# (from the file where a function is call => to the file were defined)