Generate Regular Expression from a Character Vector

This function takes a vector of strings and returns a concatenated regular expression pattern. It allows adding a prefix and/or suffix to each element and escaping special characters.

Usage

get_regex_from_vector(
  vector,
  prefix_for_regex = "",
  suffix_for_regex = "",
  fix_escaping = TRUE,
  ...
)

Arguments

vector

A character vector containing the elements to convert into a regular expression.

prefix_for_regex

A string to prepend to each element of the vector in the regex pattern.

suffix_for_regex

A string to append to each element of the vector in the regex pattern.

fix_escaping

A boolean, indicating if escaping should be fixed. Default is TRUE

...

Passed to fix_escaping

Value

A single character string containing the regular expression pattern created by concatenating all elements from the input vector, with the specified prefix and suffix applied, and special characters escaped.

Details

such as a list of special_chars_to_escape, a number of escaping char. to add (default is num_escapes = 2), the escaping char. to consider, etc.

See also

Examples

# Example 1: Create a regex pattern for a vector of words
words <- c("apple", "banana", "cherry")
rgx <- get_regex_from_vector(words, "^", "$")

# Example 2: Create a regex pattern for words with escaped parentheses
words <- c("foo(bar)", "baz(qux)")
rgx <- get_regex_from_vector(words, "(", ")?", c("(", ")"), fix_escaping = TRUE)
#here we don't escape our custom pattern "(" and ")?" even if escaping of other will be fixed