Get Text Network from folder path(s) and/or github repos

This function read files in a specified folder (default to only .R files) extracting and filtering text based on a set of criteria.

Usage

get_text_network_from_project(
  folder_path = NULL,
  repos = NULL,
  first_match_to_exclude = NULL,
  prefix_for_regex_from_the_text = "\\b",
  suffix_for_regex_from_the_text = "\\(",
  filter_2nd_match_unmatched_lines = TRUE,
  filter_first_match_results = TRUE,
  filter_egolink_within_a_file = TRUE,
  ...,
  file_path_from_colname = "from",
  file_path_to_colname = "to",
  match1_colname = "first_match",
  match2_colname = "second_match",
  line_number_match2_colname = "line_number",
  content_match1_col_name = "content_match_1",
  content_match2_col_name = "content_match_2"
)

Arguments

folder_path

character A string or list representing the path(s) and/or url of local folders path to read.

repos

character A string or list representing the name(s) of github repos (e.g., 'tidyverse/stringr').

first_match_to_exclude

character A vector of strings to exclude specific results from the original matches. For example, you can exclude results like "server".

prefix_for_regex_from_the_text

character A string representing the prefix to add to each 1st match that will be turned into a new regular expressions. The default is an empty string.

suffix_for_regex_from_the_text

character A string representing a regex to add as a suffix of each match, in order to have a complete regular expression. The default is an empty string.

filter_2nd_match_unmatched_lines

logical Default = TRUE A logical value indicating whether to keep only lines that match with the second pattern if set to TRUE, or keep all lines if set to FALSE.

filter_first_match_results

logical, default = TRUE A logical value indicating whether to apply the filter for the first match. The default is TRUE.

filter_egolink_within_a_file

logical Default = TRUE. A logical value indicating whether to filter results based on "ego links" (a document referring to itself)

...

Parameters passed to construct_corpus() and srch_pattern_in_files_get_df(). These parameters are

characters values, e.g., adjusting yourself the "pattern" parameter that will be matched

logical values, e.g., "keep_comments = T"

integer values such as ignore_match_less_than_nchar, etc.

file_path_from_colname

character Default = 'from' The column name (as a string) representing the "from" file path in the result data frame.

file_path_to_colname

character Default = 'to' The column name (as a string) representing the "to" file path in the result data frame.

match1_colname

character Default = 'first_match' The column name (as a string) for the first match in the result data frame.

match2_colname

character Default = 'second_match' The column name (as a string) for the second match in the result data frame.

line_number_match2_colname

character Default = 'line_number' The column name (as a string) for the line number of the second match in the result data frame.

content_match1_col_name

character Default = 'content_match_1' The column name (as a string) for the full content of the first match.

content_match2_col_name

character Default = 'content_match_2' The column name (as a string) for the full content of the second match.

Value

A data frame wich is the edgelist of the citations network columns 'from' and 'to' indicates the files paths or urls of the matched contents.

Details

It is designed to generate a network of text by cascading text research :

1st the func' try to match a pattern within the files, optionnaly filter the initial results.

Then, a regex-pattern is constructed by appending all the 1st match

and a 2nd research extract this pattern, and construct a network (document with the 2nd match => document with the 1st match)

See also

Examples

# Example with url from github
result <- get_text_network_from_project( folder_path =  "~" )
#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |===                                                                   |   5%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |=====                                                                 |   8%  |                                                                              |======                                                                |   8%  |                                                                              |======                                                                |   9%  |                                                                              |=======                                                               |  10%  |                                                                              |=======                                                               |  11%  |                                                                              |========                                                              |  11%  |                                                                              |========                                                              |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |==========                                                            |  14%  |                                                                              |==========                                                            |  15%  |                                                                              |===========                                                           |  15%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  17%  |                                                                              |=============                                                         |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  19%  |                                                                              |==============                                                        |  20%  |                                                                              |===============                                                       |  21%  |                                                                              |===============                                                       |  22%  |                                                                              |================                                                      |  22%  |                                                                              |================                                                      |  23%  |                                                                              |=================                                                     |  24%  |                                                                              |=================                                                     |  25%  |                                                                              |==================                                                    |  25%  |                                                                              |==================                                                    |  26%  |                                                                              |===================                                                   |  27%  |                                                                              |===================                                                   |  28%  |                                                                              |====================                                                  |  28%  |                                                                              |====================                                                  |  29%  |                                                                              |=====================                                                 |  30%  |                                                                              |=====================                                                 |  31%  |                                                                              |======================                                                |  31%  |                                                                              |======================                                                |  32%  |                                                                              |=======================                                               |  33%  |                                                                              |========================                                              |  34%  |                                                                              |========================                                              |  35%  |                                                                              |=========================                                             |  35%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  37%  |                                                                              |===========================                                           |  38%  |                                                                              |===========================                                           |  39%  |                                                                              |============================                                          |  39%  |                                                                              |============================                                          |  40%  |                                                                              |=============================                                         |  41%  |                                                                              |=============================                                         |  42%  |                                                                              |==============================                                        |  42%  |                                                                              |==============================                                        |  43%  |                                                                              |===============================                                       |  44%  |                                                                              |===============================                                       |  45%  |                                                                              |================================                                      |  45%  |                                                                              |================================                                      |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |=================================                                     |  48%  |                                                                              |==================================                                    |  48%  |                                                                              |==================================                                    |  49%  |                                                                              |===================================                                   |  50%  |                                                                              |====================================                                  |  51%  |                                                                              |====================================                                  |  52%  |                                                                              |=====================================                                 |  52%  |                                                                              |=====================================                                 |  53%  |                                                                              |======================================                                |  54%  |                                                                              |======================================                                |  55%  |                                                                              |=======================================                               |  55%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  57%  |                                                                              |========================================                              |  58%  |                                                                              |=========================================                             |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |==========================================                            |  61%  |                                                                              |===========================================                           |  61%  |                                                                              |===========================================                           |  62%  |                                                                              |============================================                          |  63%  |                                                                              |=============================================                         |  64%  |                                                                              |=============================================                         |  65%  |                                                                              |==============================================                        |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |================================================                      |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |==================================================                    |  71%  |                                                                              |==================================================                    |  72%  |                                                                              |===================================================                   |  72%  |                                                                              |===================================================                   |  73%  |                                                                              |====================================================                  |  74%  |                                                                              |====================================================                  |  75%  |                                                                              |=====================================================                 |  75%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  77%  |                                                                              |======================================================                |  78%  |                                                                              |=======================================================               |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  80%  |                                                                              |========================================================              |  81%  |                                                                              |=========================================================             |  81%  |                                                                              |=========================================================             |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |===========================================================           |  84%  |                                                                              |===========================================================           |  85%  |                                                                              |============================================================          |  85%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  87%
#> Error in sub(re, "", x, perl = TRUE): input string 8 is invalid UTF-8

# Will return a network of functions
# (from the file where a function is call => to the file were defined)