Read text file with readLines and answer a df

This function reads the content of a text file line by line and stores it in a data frame. You must pick up the FIRST col' for the file_path readed, and the LAST col for the content readed intermediar are additionnal info, i.e. line_number

Usage

readlines_in_df(
  files_path,
  return_lowered_text = FALSE,
  keep_comments = FALSE,
  .verbose = TRUE,
  file_path_col_name = "file_path",
  content_col_name = "content",
  line_number_col_name = "line_number",
  char_comment = "#"
)

Arguments

files_path

character Path(s) or url(s) to the text files to read.

return_lowered_text

logical, default = FALSE TRUE for lowercasing the returned text. FALSE (the default) will preserve the readed text

keep_comments

logical, default = FALSE If FALSE - the default, lines starting with # are treated as keep_comments and ignored.

.verbose

logical, default = TRUE If TRUE, shows a progress bar while reading the file.

file_path_col_name

character, default = "file_path" Column name for the file path in the output dataframe.

content_col_name

character, default = "content" Column name for the file content in the output dataframe.

line_number_col_name

character, default = "line_number" Column name for the line numbers in the output dataframe.

char_comment

character, default = "^#" Regex for considering that a line is a comment (default is '#' at the begining of a line of the files)

Value

RETURN : data.frame with 3 col'. First col' (file_path by default) is the file_path (readed), then line_number (by default). The last column (content by default) contain the readed lines from the file.

file_path

Path of the file where the match was found.

line_number

Line number in the file.

content

Full content of the matched line.

matches

Extracted function name or matched pattern.

Examples

if (FALSE) { # \dontrun{
  df <- readlines_in_df("path/to/file.R", case.sensitive = TRUE)
  contents_readed <- df[[ncol(df)]]
  file_path_readed <-  df[[1]]
} # }