Separate commented and non-commented lines from text

This function extracts and separates commented and non-commented parts from a given vector of texts. It uses a specified delimiter pair to identify comments and returns a data frame with the full text, extracted comments, and remaining code lines.

Usage

separate_commented_lines(texts, delim_pair = c(`/*` = "*/"), .verbose = T)

Arguments

texts

character vector containing text with possible comments.

delim_pair

character A named character vector of length 1, where the name is the opening delimiter (e.g., "/*") and the value is the closing delimiter (e.g., "*/").

.verbose

logical, default = TRUE Return a progress bar if set to TRUE

Value

A data frame with three columns:

text

The original text.

comments

The extracted comments, concatenated together if multiple are found.

codelines

The remaining text after removing comments.

Examples

texts <- c(
  "Code snippet /* first comment */ and more code /* second comment */",
  "Another line /* only a comment */",
  "A line with no comments"
)

separate_commented_lines(texts, delim_pair = c("/*" = "*/"), .verbose = FALSE)
#>                                                                  text
#> 1 Code snippet /* first comment */ and more code /* second comment */
#> 2                                   Another line /* only a comment */
#> 3                                             A line with no comments
#>                                  comments                    codelines
#> 1 /* first comment *//* second comment */ Code snippet  and more code 
#> 2                    /* only a comment */                Another line 
#> 3                                    <NA>      A line with no comments