Read some files and answer a df, optionnaly try to match a text pattern e.g., detect a file with a func' defined with 'function_name <- function'

| A] Read a bunch of files and answer the content readed in a df | B] Then try to extract a pattern and return the extracted text (no match on commented lines by default) In the 'matched text' part of the result, NA meaning 'no match' | C] Finally answer a df with all the readed content Regarding the returned df : the first col' is the file path of all the matched files (first regex passed to list.files) Last col is the extracted text : this will be the prefix passed by the user, only when there is a complete match

Usage

srch_pattern_in_files_get_df(
  files_path = NULL,
  pattern = "(^| \\.|\\b)([\\.A-Za-z0-9_]+)(?=\\s*(?:<-)\\s*function)",
  match_to_exclude = NULL,
  ignore_match_less_than_nchar = 3,
  file_path_col_name = "file_path",
  content_col_name = "content",
  line_number_col_name = "line_number",
  extracted_txt_col_name = "matches",
  ...
)

Arguments

files_path

character A vector of files path path and/or url.

pattern

character, default = "\\b([A-Za-z0-9_]+)(?=\\s*(?:<-|=)\\s*(?:function|$))" A regex for matching lines and extract text. Use the regex for finding a line by extracting text

match_to_exclude

character A vector of values that will not be returned such as a match. The rows where the values match any element in this vector will be removed.

ignore_match_less_than_nchar

double, default = 2 Excluding match depending on char. number of the matched text (strictly inferior) Default exclude match of 1 char such as 'x'.

file_path_col_name

character, default = "file_path" Column name for the file path in the output dataframe (first col' of the returned df)

content_col_name

character, default = "content" Column name for the file content in the output dataframe.

line_number_col_name

character, default = "line_number" Column name for the line numbers in the output dataframe.

extracted_txt_col_name

character, default = "matches" Column name for the extracted text (last col' of the returned df)

...

Additional arguments passed to readlines_in_df For example, keep_comments (logical, default = FALSE) for taking into account the commented lines of the files

Value

A data.frame with 4 col' : first (file_path by default) contain the file_path, then line_number (by default) contain line_number, third column (content by default) containing the readed lines from the file and the LAST ONE contain the matched text, according to the regex provided by the user

See also

Examples

#Analysing the func of the package, assuming you have installed it :
pkg_path <- list.files("~", pattern = ".R$",  recursive = TRUE , full.names = TRUE  )
lines_readed <- srch_pattern_in_files_get_df(pkg_path, .verbose = FALSE)
#> Error in sub(re, "", x, perl = TRUE): input string 8 is invalid UTF-8
# Return : a dataframe of links, according to - default - pattern