Convert an igraph object to a networkD3 visualization

This function transforms an igraph graph object into a networkD3 interactive visualization. It allows customization of titles, node colors, missing link colors, and additional HTML code insertion.

Usage

get_networkd3_from_igraph(
  graph_igraph,
  title_h1 = "networkD3",
  subtitle_h2 = NULL,
  endnotes_h3 = NULL,
  colors_for_nodes = c("green", "grey", "black", "red"),
  color_outdeg_instead_of_indeg = F,
  color_for_na_link = colors_for_nodes[[1]],
  add_html_div_code_before_graph = NULL,
  ...,
  charge = -200
)

Arguments

graph_igraph

An igraph object representing the network.

title_h1

Character. The main title (HTML <h1>) displayed above the graph. Default is "networkD3".

subtitle_h2

Character (optional). The subtitle (HTML <h2>) displayed below the main title. Default is NULL (no subtitle).

endnotes_h3

Character (optional). A description or note (HTML <h3>) displayed after the network. Default is NULL (no text).

colors_for_nodes

Character vector. Specifies colors for nodes based on their degree and also number of separation the func' will try to realize (based on indegrees or outdegrees). Default is c("green", "grey", "black", "red").

color_outdeg_instead_of_indeg

Logical. If TRUE, node colors are assigned based on out-degree instead of in-degree. Default is FALSE.

color_for_na_link

Character. Color assigned to links when data is missing. Default is the 1st color of colors_for_node ("green").

add_html_div_code_before_graph

Character (optional). Custom HTML code inserted after the end of the network. Default is NULL (no additional HTML).

...

optional parameter passed to networkD3::forceNetwork

charge

integer, default = -200 Numeric value indicating either the strength of the node repulsion (negative value) or attraction (positive value). Passed to networkD3::forceNetwork.

Value

A list with the net3d usual object from networkD3:: and the html object from networkD3::forceNetwork :

net3d

networkD3 usual list

$links

$links is the 1st entry of a networkD3 usual list with 'source' and 'target' vectors

$nodes

$nodes is the 2nd entry of a networkD3 usual list, with the 'name' of the nodes vector

forcenetwork

networkD3::forceNetwork usual html object for interactive network visualization

Examples

library(igraph)
#> 
#> Attaching package: ‘igraph’
#> The following objects are masked from ‘package:stats’:
#> 
#>     decompose, spectrum
#> The following object is masked from ‘package:base’:
#> 
#>     union
g <- make_ring(10)
get_networkd3_from_igraph(g, title_h1 = "My Network")
#> $net3d
#> $net3d$links
#>    source target color_deg
#> 1       0      1   #FF0000
#> 2       1      2   #FF0000
#> 3       2      3   #FF0000
#> 4       3      4   #FF0000
#> 5       4      5   #FF0000
#> 6       5      6   #FF0000
#> 7       6      7   #FF0000
#> 8       7      8   #FF0000
#> 9       0      9   #FF0000
#> 10      8      9   #FF0000
#> 
#> $net3d$nodes
#>    name indegree outdegree color_deg
#> 1     1        2         2   #FF0000
#> 2     2        2         2   #FF0000
#> 3     3        2         2   #FF0000
#> 4     4        2         2   #FF0000
#> 5     5        2         2   #FF0000
#> 6     6        2         2   #FF0000
#> 7     7        2         2   #FF0000
#> 8     8        2         2   #FF0000
#> 9     9        2         2   #FF0000
#> 10   10        2         2   #FF0000
#> 
#> 
#> $forcenetwork
#>